Moving one more script that was misplaced
